requirements.txt
setuptools
numpy
scipy
pyserial
matplotlib
pyqtgraph
PyQt5
